# Analyze React+Redux perf timings using recording analysis