--- a/packages/shared/client/types.ts
+++ b/packages/shared/client/types.ts
@@ -63,6 +63,7 @@ import {
   getExceptionValueResult,
   getFrameArgumentsResult,
   getParentNodesResult,
+  getPointStackResult,
   getScopeResult,
   getSourceOutlineResult,
   getTopFrameResult,
@@ -239,6 +240,7 @@ export interface ReplayClientInterface {
   getParentNodes(pauseId: PauseId, nodeId: string): Promise<getParentNodesResult>;
   getPointNearTime(time: number): Promise<TimeStampedPoint>;
   getPointsBoundingTime(time: number): Promise<PointsBoundingTime>;
+  getPointStack(point: ExecutionPoint, maxCount?: number): Promise<getPointStackResult>;
   getRecordingId(): RecordingId | null;
   getNetworkRequestBody(
     requestId: RequestId,