--- a/src/ui/actions/eventListeners/jumpToCode.ts
+++ b/src/ui/actions/eventListeners/jumpToCode.ts
@@ -368,7 +368,6 @@ export const eventListenerLocationCache: Cache<
           replayClient,
           replayEventType,
           onClickPreview.preview,
-          state.sources,
           "react"
         );