This PR:

- Adds a new "React+Redux Perf" sidebar panel
- Adds a button that, when clicked:
  - Finds all Redux dispatches
  - Finds all calls to `useSelector`
  - Finds all unique selector functions passed to `useSelector`
  - Finds all individual calls to each selector function and calculates their durations
  - Sums up the total execution time for each selector function
  - Finds all calls to notify React-Redux subscribers and checks the time spent notifying subscribers
  - Finds all calls to re-render a React component
  - Buckets the data by `dispatch()` start/end times, and groups all the selector calls, notification calls, and queued re-renders triggered by that dispatch

In other words, this is a major first step towards a useful React+Redux perf analysis tool.

Caveats:

- Right now I'm just logging the grouped data, and haven't tried to display or visualize any of this yet
- It currently takes ~30s to calculate the results for a 2s focus region with ~40 dispatches. The most expensive parts are running evaluations to find all the selector functions, and calling `Debugger.findStepOutTarget` to determine the execution time for a couple thousand function calls.

It's very likely that the execution speed would be faster if we moved this into a backend postprocessing routine, and also added some additional backend APIs to allow requesting step-out results in batches instead of individual requests.

Here's an example of what the output looks like for a given dispatch in our own front-end codebase:

![image](https://github.com/replayio/devtools/assets/1128784/6c35b87a-dccc-4c9d-87a4-c06299b2d2e1)

That one dispatch:

- did queue a re-render (by scheduling one component to update)
- we spent 15.3ms notifying subscribers
- we cascaded through 10 nested `connect` components notifying their children
- we ran 114 separate selector calls
- of those, only 14 selector calls had a meaningful non-zero duration
- total time of selectors was 0.3ms, which means that there's 15.0ms of subscriber overhead (!)