@ryota-murakami 2023-06-29
@markerikson [packages/replay-next](https://github.com/replayio/devtools/pull/9410/files#diff-50db83862d89c7c2c093e7f276ff579f064d404ec8d75d8cbed973babad64dcb) meant Next.js exclusive package?