@markerikson 2023-06-30
@ryota-murakami : no. All our app source code originally lived in the `src` folder.  When Brian began rewriting some of the pieces using newer React techniques like Suspense, he created a `packages/bvaughn-architecture` folder and put the new code there. Later, we renamed it to `packages/replay-next`, to indicate that it's "the next generation of Replay app source code".