@markerikson 2023-06-30
I just updated this to capture details on the start and end of the related React commit, plus counts of how many function and class components got rendered overall:

![image](https://github.com/replayio/devtools/assets/1128784/cccac568-b065-4030-8bba-568d9d90f332)

as well as inspecting the rendered components and determining the details for each unique component type:

![image](https://github.com/replayio/devtools/assets/1128784/a83ad1cf-ec22-4cb6-81c1-387fba89344e)


This gives us a foundation for being able to do real-world React render performance analysis, on _production_ apps, no further profiling setup needed!