@markerikson 2023-07-06
Keep it as a draft for now. Like the other "React Render" panel, it's something I intend to merge and feature flag, but I need to do some cleanup before it's ready to at least have in the codebase.