@markerikson 2023-09-11
There's a lot of conceptual overlap between what I was doing in the first iteration of this "Redux+React Perf Profiler" PR, and what I did in [the original POC for the "React Render Timeline" panel](https://github.com/replayio/devtools/pull/8915).

Both rely on fetching hit points for internals. Both require fetching points for the current focus window. And my POC for the Redux perf work wanted to know when React had rendered, and when we _queued_ renders, so that we could know if the dispatch did actually result in components rendering.

Over this weekend, I've updated this PR:

- Rebased against `main`
- Extracted the basic Redux `dispatch()` checks and timings as an interval cache
- Made improvements to the function detail formatting logic
- And tonight I rewrote the React Render Timeline to also use an interval cache, use the new `Session.getPointStack` API for stack frames, and rewrote the items as a virtualized list.

The result is that the React Render Timeline is showing up _significantly_ faster (5-10s for several hundred hits), whereas before it would take seemingly _forever_ for values to slowly come in.

I don't have any UI for the Redux data yet, but that should be trivial to add now that I've got the cache for the values.

From there, the next step would be to start adding logic to query for perf stats related to a given dispatch: was a render queued, how long did it take, how many components rendered, how long did _those_ take, which ones were most expensive, and what selectors ran and _their_ timings.  Rather than fetch all of that up front, fetching it for just one render ought to be reasonably faster.

Here's what the render timeline looks like atm:

![image](https://github.com/replayio/devtools/assets/1128784/7622e5a0-b960-4ffe-818e-0b06611c99a9)