@markerikson 2023-09-12
I've finally built out some UI for the Redux Dispatch Perf panel:

![image](https://github.com/replayio/devtools/assets/1128784/7777b431-97bc-4355-aadd-aa44d4e556bf)

things I can now do with this info:

- set focus ranges from the timestamps
- determine if any React renders were queued by the notifications
- fetch info about selectors, although that gets trickier because I really need to know what selectors were in use _prior_ to the dispatch

and similarly, with React renders, I could pull together info on:

- which components rendered
- how long those each took

I _know_ there's all kinds of caveats on sync renders vs concurrent rendering, components being thrown away, etc. Not _anywhere_ close to being a "real" profiler.  But I've got demo code (even if commented out atm) that shows I can get plausible interpretations of that info.  So it's a starting point.