@markerikson 2023-09-15
Tonight's progress: it's finding the likely start and end time of the next sync render right after the dispatch, and displaying that duration:

![image](https://github.com/replayio/devtools/assets/1128784/ea7d4e09-8f5e-42fc-838d-4102eea7b0b9)

Also have it displaying how many components _probably_ were triggered to render as a direct result of the dispatch.

(Need to start extracting details on those components to confirm, but the poking around I've done says it's plausible)