@markerikson 2023-09-20
Talked with Jon. He likes the general direction of the UI here, including reusing the "details pane locked to the bottom of the panel" approach that we're using on the Protocol Viewer and Test Suites panels.

We'll keep these two panels separate for now, but I'll update the Perf panel to have a Redux icon.

Longer-term, there's a lot of things we _could_ do here (maybe someday having a "popout" button that shows the list in a table format in the area where our source viewer currently lives).