@markerikson 2023-09-21
I'm going to call this sufficiently ready that I'd like to get it merged in, behind the existing `feature_reactPanel` experimental flag.



### Current functionality

- React Queued Renders Panel
  - Fetches hits to `scheduleUpdateOnFiber` using an interval cache
  - Fetches stack frames via `getPointStack`, filters via URLs to find a likely candidate call to `setState()` or similar
  - Renders a virtualized list of function names, and has Jump to Code
- Redux Dispatch Perf Panel
  - Fetches hits to Redux `dispatch()` internals
  - Finds all likely Redux dispatches and extracts the action types (even if the Redux DevTools were not enabled)
  - Renders a virtualized list of action types, and has Jump to Code (yes this _highly_ overlaps with our actual Redux DevTools functionality - we'll sort that out later)
  - Calculates durations of the reducer, subscriber notifications, and total dispatch time
  - Uses `scheduleUpdateOnFiber` to estimate how many React components were likely queued to render by this dispatch (by checking `sUOF` calls inside of the subscriber notification start/end)
  - Tries to find a likely React render caused by this dispatch, and if found, displays the estimated total duration of that render

Both panels use Suspense to load, and `react-window` for the virtualized lists.  

### Screenshots

React Render Panel:

![image](https://github.com/replayio/devtools/assets/1128784/2f8840cc-5a4a-4977-8339-d28cfcc26fec)


Redux Perf Panel:

![image](https://github.com/replayio/devtools/assets/1128784/61db3cbf-3887-4e93-9c6a-b6a5ff484adc)


### Future Work

There's a lot of things I've proved that I _can_ do that are not in the current version of this PR:

- Counting how many function components and class components rendered in a given render pass
- Finding all Redux selectors used in an app
- Counting how many selectors got called during a set of subscriber notifications, and calculating their durations
- Showing list items for React render commits in the React panel (ie, alternating calls to queue a render vs commits)
- Showing the red "you are here" line in the lists


What I've got so far is good enough to at least have in the codebase behind a flag, and we can build on these from here.