@bvaughn 2023-09-22
Let's chat about this PR and the topic in person at the offsite this coming week?