@markerikson 2023-10-12
Superseded by #9775 .  The Redux perf half went into #9776 , but we've decided not to merge that into the codebase for now.  I'll rebase it, then close the PR.